+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ## Unreleased
 
+### Enhancements
+- Added SQL Server 2016 support for Query Performance Monitoring when using DMV-only mode (QUERY_MONITORING_DISABLE_HISTORICAL_INFORMATION=true)
+
 ## v2.25.2 - 2025-12-09
 
 ### ⛓️ Dependencies